Added delete function for a card

+++ app/courses/index/route.js
@@ -2,13 +2,14 @@ import Ember from 'ember';
 
 export default Ember.Route.extend({
   model: function() {
-    // Uncomment this when the API is ready
     return this.store.findAll('course');
   },
 
   actions: {
     deleteCourse: function(course) {
-      course.destroyRecord();
+      if (window.confirm('Are you sure you want to delete?')) {
+        course.destroyRecord();
     }
   }
+}
 });
